Move morency doin material model to cookbook
