+++ cookbooks/morency_doin_2004/morency_doin.prm
