+++ doc/manual/cookbooks/morency_doin/morency_doin.part.prm
@@ -1,0 +1,59 @@
+subsection Geometry model
+  set Model name = box
+
+  subsection Box
+    set X extent      = 3000e3
+    set Y extent      = 750e3
+    set X repetitions = 4
+  end
+end
+
+subsection Compositional fields
+  set Number of fields = 2
+  set Names of fields  = upper_crust, lower_crust
+end
+
+subsection Initial composition model
+  set Model name = function
+
+  subsection Function
+    set Variable names = x,y
+    set Function expression = if(y>=725e3,1,0);if((y<725e3&y>700e3),1,0)
+  end
+end
+
+subsection Initial temperature model
+  set Model name = function
+
+  subsection Function
+    set Variable names = x,y
+    set Function constants = h=750e3, w=3000e3, mantleT=1350 # deg C
+    set Function expression = \
+      if( y < 100e3, \
+        (100e3-y)/100e3*(1600-mantleT)+mantleT+293, \
+        if(y>650e3, \
+          (h-y)/(100e3)*mantleT+293, \
+          mantleT+293))
+  end
+end
+
+subsection Material model
+  set Model name = Morency and Doin
+
+  subsection Morency and Doin
+    set Densities = 3300,2920,2920
+    set Activation energies = 500,320,320
+    set Coefficient of yield stress increase with depth = 0.25
+    set Thermal expansivities = 3.5e-5
+    set Stress exponents for viscous rheology = 3
+    set Stress exponents for plastic rheology = 30
+    set Thermal diffusivity = 0.8e-6
+    set Heat capacity = 1.25e3
+    set Activation volume = 6.4e-6
+    set Reference strain rate = 6.4e-16
+    set Preexponential constant for viscous rheology law = 7e11 ## Value used in paper is 1.24e14
+    set Cohesive strength of rocks at the surface = 117
+    set Reference temperature = 293
+    set Minimum strain rate = 5e-19                             ## Value used in paper is 1.4e-20
+  end
+end
