+++ tests/morency_doin.cc
@@ -1,4 +1,2 @@
-#include "material_model_dependencies.cc"
 #include "../cookbooks/morency_doin_2004/morency_doin.h"
 #include "../cookbooks/morency_doin_2004/morency_doin.cc"
-
